requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
PyYAML==6.0.2
pyzmq==26.2.0
